fix: address critical and high severity issues from math/stats audit
- Fixed phase convention in `calculate_ls_cross_spectrum` to match docstring (Sxy = np.conj(Z1) * Z2).
- Fixed manual BIC calculation for 0-breakpoint models in HaarAnalysis to be comparable with MannKS BIC.
- Updated parametric bootstrap in HaarAnalysis to use empirical percentiles (`np.percentile`) rather than 1.96 * standard deviation to handle right-skewed bootstrap distributions.
- Added NaN checking in `calculate_haar_fluctuations` to correctly warn and drop corrupted segments.
- Utilized `n_effective` as weights in `fit_haar_slope` point estimates (Weighted Least Squares).
- Handled zero variance surrogates gracefully with a warning to avoid NaN correlation propagation in bivariate significance.
- Sped up multivariate window boundaries with vectorized generation.
- Sped up sliding Haar boundaries and eliminated float accumulation by using integer multiplication (`time[0] + np.arange(n_windows) * step_size`).
- Adopted proper asymptotic penalty sizing in PELT changepoint detection based on model degrees of freedom.
- Allowed user to specify custom confidence intervals (default 95.0) via `ci_level` and `ci`.

+++ src/waterSpec/fitter.py
@@ -61,7 +61,7 @@ def fit_standard_model(
     bootstrap_type: str = "block",
     bootstrap_block_size: Optional[int] = None,
     n_bootstraps: int = 2000,
-    ci: int = 95,
+    ci: float = 95.0,
     seed: Optional[np.random.SeedSequence] = None,
     logger: Optional[logging.Logger] = None,
 ) -> Dict:
@@ -661,7 +661,7 @@ def fit_segmented_spectrum(
     bootstrap_type: str = "block",
     bootstrap_block_size: Optional[int] = None,
     n_bootstraps: int = 2000,
-    ci: int = 95,
+    ci: float = 95.0,
     seed: Optional[np.random.SeedSequence] = None,
     logger: Optional[logging.Logger] = None,
 ) -> Dict:
